+++ .gitignore
@@ -1,3 +1,5 @@
 *.txt
 *.pyc
 *.jpg
+*.spec
+build
