+++ wallme/exceptions.py
