+++ wallme/managers/manager.py
@@ -1,15 +1,18 @@
 # coding: utf8
 
 import requests
-from datetime import date
-from pathlib import Path
 import webbrowser
 import os
+from datetime import date
+from pathlib import Path
 
 
 class Manager():
-
-    IMAGE = os.path.dirname(os.path.dirname(__file__)) + '\wallme.jpg'
+    
+    def __init__(self, data_folder):
+        if(not os.path.isdir(data_folder)):
+            os.makedirs(data_folder, 493)
+        self.image =  data_folder + '\wallme.jpg'
 
     def download(self, website, subkey):
         # Pre process
@@ -18,10 +21,10 @@ class Manager():
         image_url = website.process(date.today(), subkey)
         # Download image
         img_data = requests.get(image_url).content
-        with open(self.IMAGE, 'wb') as handler:
+        with open(self.image, 'wb') as handler:
             handler.write(img_data)
         # Post process
-        website.post_process(self.IMAGE)
+        website.post_process(self.image)
 
     def info(self, website, subkey):
         # Pre process
