+++ wallme/managers/managerfactory.py
