+++ wallme/managers/windows.py
@@ -4,11 +4,15 @@ import ctypes
 import winreg
 from .manager import Manager
 from ..exceptions import WallmeException
-
+import os
 
 class Windows(Manager):
 
     REGISTRY_KEY = "wallme"
+    DATA_FOLDER = os.path.expandvars(r'%LOCALAPPDATA%\Wallme')
+
+    def __init__(self):
+        super().__init__(self.DATA_FOLDER)
 
     def set(self, website, subkey, test=False):
         super().download(website, subkey)
